// projects/ngrx-app-store/src/lib/actions/entity-actions.enum.ts
export enum BaseEntityActions {
  DefaultLoad = '[Default] LOAD REQUESTED',
  DefaultLoadSuccess = '[Default] LOAD SUCCESS',
  DefaultCreate = '[Default] CREATE REQUESTED',
  DefaultCreateSuccess = '[Default] CREATE SUCCESS',
  DefaultUpdate = '[Default]  UPDATE REQUESTED',
  DefaultUpdateSuccess = '[Default] UPDATE SUCCESS',
  DefaultRemove = '[Default] REMOVE REQUESTED',
  DefaultRemoveSuccess = '[Default] REMOVE SUCCESS',
  LoadUsers = '[Users] Load',
  LoadUsersComplete = '[Users] Load:Complete',
  DefaultError = '[Default] Error'

}